account-type-discount decimale

+++ api/src/main/java/ca/ids/abms/modules/accounts/AccountService.java
@@ -370,7 +370,7 @@ public class AccountService {
  
 
     private void validateAccountTypeDiscount(Account account) {
-    Integer accountTypeDiscount = account.getAccountTypeDiscount();
+    Double accountTypeDiscount = account.getAccountTypeDiscount();
     if (accountTypeDiscount ==null ) {
     return ;
     }
